Seismograph Visual Improvements
Added offsets to the seismograph that keeps the curve's writing point locked to simulate 'printing' the wave. Still need to finetune the ability to dynamically resize / change the animation speed and time of the curve.
Implemented basic Citizen class in anticipation of creating the Manual and Potential Offerings mechanic.
To-do next: need to make some more indepth mockups of various interfaces and continue art direction.